refactor: rename folder `admin` -> ``Protected`
